Refactor: 디렉토리 구조를 간결하게 바꾼다
