+++ src/main/java/Skeep/backend/location/location/domain/LocationRepository.java
@@ -1,4 +1,4 @@
-package Skeep.backend.location.domain;
+package Skeep.backend.location.location.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
