+++ src/main/java/Skeep/backend/location/userLocation/domain/UserLocationRepository.java
@@ -1,4 +1,4 @@
-package Skeep.backend.location.domain;
+package Skeep.backend.location.userLocation.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
