+++ src/main/java/Skeep/backend/location/userLocation/service/UserLocationSaver.java
@@ -1,7 +1,7 @@
-package Skeep.backend.location.service;
+package Skeep.backend.location.userLocation.service;
 
-import Skeep.backend.location.domain.UserLocation;
-import Skeep.backend.location.domain.UserLocationRepository;
+import Skeep.backend.location.userLocation.domain.UserLocation;
+import Skeep.backend.location.userLocation.domain.UserLocationRepository;
 import Skeep.backend.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
