+++ src/main/java/Skeep/backend/screenshot/service/ScreenshotService.java
@@ -5,11 +5,10 @@ import Skeep.backend.global.exception.BaseException;
 import Skeep.backend.gpt.service.GptService;
 import Skeep.backend.kakaoMap.dto.response.KakaoResponseResult;
 import Skeep.backend.kakaoMap.service.KakaoMapService;
-import Skeep.backend.location.domain.Location;
-import Skeep.backend.location.domain.LocationRepository;
-import Skeep.backend.location.domain.UserLocation;
-import Skeep.backend.location.service.LocationService;
-import Skeep.backend.location.service.UserLocationSaver;
+import Skeep.backend.location.location.domain.Location;
+import Skeep.backend.location.userLocation.domain.UserLocation;
+import Skeep.backend.location.location.service.LocationService;
+import Skeep.backend.location.userLocation.service.UserLocationSaver;
 import Skeep.backend.naverOcr.service.NaverOcrService;
 import Skeep.backend.s3.service.S3Service;
 import Skeep.backend.screenshot.dto.request.ScreenshotUploadDto;
