+++ src/test/java/Skeep/backend/location/location/domain/LocationTest.java
@@ -1,4 +1,4 @@
-package Skeep.backend.location.domain;
+package Skeep.backend.location.location.domain;
 
 import Skeep.backend.category.domain.ECategory;
 import org.junit.jupiter.api.DisplayName;
